move esb related templates from core to esb
