+++ plugins/org.wso2.developerstudio.eclipse.artifact.proxyservice/src/org/wso2/developerstudio/eclipse/artifact/proxyservice/utils/psartifactconstants.properties
@@ -2,7 +2,7 @@ PS_WIZARD_WINDOW_TITLE=New Proxy Service
 PASS_THROUGH_PROXY=Pass Through Proxy
 SECURE_PROXY=Secure Proxy
 WSDL_BASED_PROXY=WSDL Based Proxy
-LOGGING_PROXY=Log Forward Proxy
+LOGGING_PROXY=Logging Proxy
 TRANSFORMER_PROXY=Transformer Proxy
 CUSTOM_PROXY=Custom Proxy
 ESB_PROJECT_NATURE=org.wso2.developerstudio.eclipse.esb.project.nature
