+++ plugins/org.wso2.developerstudio.eclipse.artifact.template.sequence/META-INF/MANIFEST.MF
@@ -7,7 +7,8 @@ Bundle-Activator: org.wso2.developerstudio.eclipse.artifact.template.Activator
 Require-Bundle: org.wso2.developerstudio.eclipse.kernel.libraries;bundle-version="[4.1.0,7.0.0)",
  org.wso2.developerstudio.eclipse.platform.core;bundle-version="[4.1.0,7.0.0)",
  org.wso2.developerstudio.eclipse.platform.ui;bundle-version="[4.1.0,7.0.0)",
- org.wso2.developerstudio.eclipse.esb.libraries;bundle-version="[6.0.0,7.0.0)"
+ org.wso2.developerstudio.eclipse.esb.libraries;bundle-version="[6.0.0,7.0.0)",
+ org.wso2.developerstudio.eclipse.project.extensions;bundle-version="[6.3.0,7.0.0)"
 Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
 Import-Package: org.eclipse.core.resources,
